Move code in frame.h and rigid_body.h to frame.cc and rigid_body.cc to speed subsequent PRs.
