+++ multibody/tree/rigid_body.cc
@@ -139,6 +139,41 @@ void RigidBody<T>::SetCenterOfMassInBodyFrameAndPreserveCentralInertia(
   SetCenterOfMassInBodyFrameNoModifyInertia(context, pf_BoBcm_B);
 }
 
+template <typename T>
+void RigidBody<T>::AddInForce(const systems::Context<T>& context,
+                              const Vector3<T>& p_BP_E,
+                              const SpatialForce<T>& F_Bp_E,
+                              const Frame<T>& frame_E,
+                              MultibodyForces<T>* forces) const {
+  DRAKE_THROW_UNLESS(forces != nullptr);
+  DRAKE_THROW_UNLESS(
+      forces->CheckHasRightSizeForModel(this->get_parent_tree()));
+  const math::RotationMatrix<T> R_WE =
+      frame_E.CalcRotationMatrixInWorld(context);
+  const Vector3<T> p_PB_W = -(R_WE * p_BP_E);
+  const SpatialForce<T> F_Bo_W = (R_WE * F_Bp_E).Shift(p_PB_W);
+  AddInForceInWorld(context, F_Bo_W, forces);
+}
+
+template <typename T>
+Vector3<T> RigidBody<T>::CalcCenterOfMassTranslationalVelocityInWorld(
+    const systems::Context<T>& context) const {
+  const RigidBody<T>& body_B = *this;
+  const Frame<T>& frame_B = body_B.body_frame();
+
+  // Form frame_B's spatial velocity in the world frame W, expressed in W.
+  const SpatialVelocity<T>& V_WBo_W =
+      body_B.EvalSpatialVelocityInWorld(context);
+
+  // Form v_WBcm_W, Bcm's translational velocity in frame W, expressed in W.
+  const Vector3<T> p_BoBcm_B = CalcCenterOfMassInBodyFrame(context);
+  const math::RotationMatrix<T> R_WB =
+      frame_B.CalcRotationMatrixInWorld(context);
+  const Vector3<T> p_BoBcm_W = R_WB * p_BoBcm_B;
+  const Vector3<T> v_WBcm_W = V_WBo_W.Shift(p_BoBcm_W).translational();
+  return v_WBcm_W;
+}
+
 }  // namespace multibody
 }  // namespace drake
