+++ multibody/tree/rigid_body.h
@@ -484,16 +484,8 @@ class RigidBody : public MultibodyElement<T> {
   void AddInForce(
       const systems::Context<T>& context,
       const Vector3<T>& p_BP_E, const SpatialForce<T>& F_Bp_E,
-      const Frame<T>& frame_E, MultibodyForces<T>* forces) const {
-    DRAKE_THROW_UNLESS(forces != nullptr);
-    DRAKE_THROW_UNLESS(
-        forces->CheckHasRightSizeForModel(this->get_parent_tree()));
-    const math::RotationMatrix<T> R_WE =
-        frame_E.CalcRotationMatrixInWorld(context);
-    const Vector3<T> p_PB_W = -(R_WE * p_BP_E);
-    const SpatialForce<T> F_Bo_W = (R_WE * F_Bp_E).Shift(p_PB_W);
-    AddInForceInWorld(context, F_Bo_W, forces);
-  }
+      const Frame<T>& frame_E, MultibodyForces<T>* forces) const;
+
   /// Gets this body's center of mass position from the given context.
   /// @param[in] context contains the state of the multibody system.
   /// @returns p_BoBcm_B position vector from Bo (this rigid body B's origin)
@@ -512,25 +504,7 @@ class RigidBody : public MultibodyElement<T> {
   /// @retval v_WBcm_W The translational velocity of Bcm (this rigid body's
   /// center of mass) in the world frame W, expressed in W.
   Vector3<T> CalcCenterOfMassTranslationalVelocityInWorld(
-      const systems::Context<T>& context) const {
-    const RigidBody<T>& body_B = *this;
-    const Frame<T>& frame_B = body_B.body_frame();
-
-    // Form frame_B's spatial velocity in the world frame W, expressed in W.
-    const SpatialVelocity<T>& V_WBo_W =
-        body_B.EvalSpatialVelocityInWorld(context);
-
-    // Form v_WBcm_W, Bcm's translational velocity in frame W, expressed in W.
-    const Vector3<T> p_BoBcm_B = CalcCenterOfMassInBodyFrame(context);
-    const math::RotationMatrix<T> R_WB =
-        frame_B.CalcRotationMatrixInWorld(context);
-    const Vector3<T> p_BoBcm_W = R_WB * p_BoBcm_B;
-    const Vector3<T> v_WBcm_W = V_WBo_W.Shift(p_BoBcm_W).translational();
-    return v_WBcm_W;
-  }
-
-  // TODO(joemasterjohn): Speed this up when we can store a reference to a
-  //  SpatialInertia<T> as an abstract parameter.
+      const systems::Context<T>& context) const;
 
   /// Gets this body's spatial inertia about its origin from the given context.
   /// @param[in] context contains the state of the multibody system.
@@ -541,6 +515,8 @@ class RigidBody : public MultibodyElement<T> {
   /// @pre the context makes sense for use by this RigidBody.
   SpatialInertia<T> CalcSpatialInertiaInBodyFrame(
       const systems::Context<T>& context) const {
+    // TODO(joemasterjohn): Speed this up when we can store a reference to a
+    //  SpatialInertia<T> as an abstract parameter.
     const systems::BasicVector<T>& spatial_inertia_parameter =
         context.get_numeric_parameter(spatial_inertia_parameter_index_);
     return internal::parameter_conversion::ToSpatialInertia(
